+++ src/seekstorm/lib.rs
@@ -28,7 +28,7 @@
 //! # tokio_test::block_on(async {
 //! use std::path::Path;
 //! use std::sync::{Arc, RwLock};
-//! use seekstorm::index::{IndexMetaObject, SimilarityType,TokenizerType,StopwordType,FrequentwordType,AccessType,StemmerType,NgramSet,create_index};
+//! use seekstorm::index::{IndexMetaObject, SimilarityType,TokenizerType,StopwordType,FrequentwordType,AccessType,StemmerType,NgramSet,DocumentCompression,create_index};
 //!
 //! let index_path=Path::new("C:/index/");
 //! let schema_json = r#"
@@ -39,12 +39,13 @@
 //! let meta = IndexMetaObject {
 //! id: 0,
 //! name: "test_index".into(),
-//! similarity:SimilarityType::Bm25f,
-//! tokenizer:TokenizerType::AsciiAlphabetic,
-//! stemmer:StemmerType::None,
+//! similarity: SimilarityType::Bm25f,
+//! tokenizer: TokenizerType::AsciiAlphabetic,
+//! stemmer: StemmerType::None,
 //! stop_words: StopwordType::None,
-//! frequent_words:FrequentwordType::English,
-//! ngram_indexing:NgramSet::NgramFF as u8,
+//! frequent_words: FrequentwordType::English,
+//! ngram_indexing: NgramSet::NgramFF as u8,
+//! document_compression:  DocumentCompression::Snappy,
 //! access_type: AccessType::Mmap,
 //! spelling_correction: None,
 //! query_completion: None,
@@ -373,7 +374,7 @@
 //! ```no_run
 //! # tokio_test::block_on(async {
 //! use std::path::Path;
-//! use seekstorm::index::{IndexMetaObject, SimilarityType,TokenizerType,StopwordType,FrequentwordType,AccessType,StemmerType,NgramSet,create_index};
+//! use seekstorm::index::{IndexMetaObject, SimilarityType,TokenizerType,StopwordType,FrequentwordType,AccessType,StemmerType,NgramSet,DocumentCompression,create_index};
 //!
 //! let index_path=Path::new("C:/index/");
 //! let schema_json = r#"
@@ -385,12 +386,13 @@
 //! let meta = IndexMetaObject {
 //!     id: 0,
 //!     name: "test_index".into(),
-//!     similarity:SimilarityType::Bm25f,
-//!     tokenizer:TokenizerType::AsciiAlphabetic,
-//!     stemmer:StemmerType::None,
+//!     similarity: SimilarityType::Bm25f,
+//!     tokenizer: TokenizerType::AsciiAlphabetic,
+//!     stemmer: StemmerType::None,
 //!     stop_words: StopwordType::None,
-//!     frequent_words:FrequentwordType::English,
-//!     ngram_indexing:NgramSet::NgramFF as u8,
+//!     frequent_words: FrequentwordType::English,
+//!     ngram_indexing: NgramSet::NgramFF as u8,
+//!     document_compression:  DocumentCompression::Snappy,
 //!     access_type: AccessType::Mmap,
 //!     spelling_correction: None,
 //!     query_completion: None,